# hardware_store/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для магазина бытовой техники",
        "title": "Hardware Store API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/images/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Возвращает бинарные данные изображения по уникальному идентификатору",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Получить изображение по ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бинарные данные изображения"
                    },
                    "400": {
                        "description": "Невалидный формат UUID изображения",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Изображение не найдено",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при получении изображения",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Обновляет существующее изображение новыми бинарными данными",
                "consumes": [
                    "application/octet-stream"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Обновить изображение",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Бинарные данные изображения",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Изображение успешно обновлено"
                    },
                    "400": {
                        "description": "Невалидный формат UUID или некорректные данные изображения",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Изображение не найдено",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при обновлении",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Удаляет изображение по уникальному идентификатору UUID",
                "tags": [
                    "images"
                ],
                "summary": "Удалить изображение",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Изображение успешно удалено"
                    },
                    "400": {
                        "description": "Невалидный формат UUID изображения",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Изображение не найдено",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при удалении",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/products/{id}/image": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Возвращает изображение продукта по его уникальному идентификатору",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Получить изображение продукта",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID продукта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бинарные данные изображения"
                    },
                    "400": {
                        "description": "Невалидный формат UUID продукта",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Изображение не найдено",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при получении изображения",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Загружает новое изображение для указанного продукта",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Загрузить изображение для продукта",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID продукта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Бинарные данные изображения",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Изображение успешно загружено",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ImageResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидный формат UUID продукта или некорректные данные изображения",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при сохранении изображения",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "hardware_store_internal_web_dto.ImageResponse": {
            "description": "Данные изображения включая уникальный идентификатор",
            "type": "object",
            "properties": {
                "image_id": {
                    "type": "string",
                    "example": "b1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
                }
            }
        },
        "hardware_store_internal_web_dto.InternalErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
        "hardware_store_internal_web_dto.NotFoundErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "client not found"
                }
            }
        },
...
        "hardware_store_internal_web_dto.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid uuid format"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: \"Bearer \u003cваш_токен\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "tokenUrl": "https://oauth2.googleapis.com/token",
            "scopes": {
                "email": "Просмотр вашего email",
                "profile": "Просмотр вашего профиля"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        },
        {
            "OAuth2": []
        }
    ]
}